// types.ts
export enum ParticipationType {
    PRESIDENTE = 'Presidente',
    ORACAO_INICIAL = 'Oração Inicial',
    ORACAO_FINAL = 'Oração Final',
    TESOUROS = 'Tesouros da Palavra de Deus',
    MINISTERIO = 'Faça Seu Melhor no Ministério',
    VIDA_CRISTA = 'Nossa Vida Cristã',
    DIRIGENTE = 'Dirigente do EBC',
    LEITOR = 'Leitor do EBC',
    AJUDANTE = 'Ajudante',
    CANTICO = 'Cântico',
    COMENTARIOS_FINAIS = 'Comentários Finais',
}